% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{bank_results}
\alias{bank_results}
\title{Access the results of Dodd-Frank Act Stress Tests}
\usage{
bank_results(
  banklevel_only = TRUE,
  consistent_sample_from_year = NULL,
  cache_file = TRUE
)
}
\arguments{
\item{banklevel_only}{Boolean. If TRUE, aggregate results are filtered out of the dataset.}

\item{consistent_sample_from_year}{Either NULL or a four-digit year starting with 2013. If a year is provided, the consistent sample (ie, banks with information for all years including and after that) is returned.}

\item{cache_file}{Boolean. If TRUE, the raw file is downloaded to a local folder and used to load the information in subsequent runs.}
}
\value{
A tidy tibble with the stress tests results.
}
\description{
Access the results of Dodd-Frank Act Stress Tests
}
